Refactor file I/O for large file support (fseek_las, ftell_las, stat_las) and update data types to prevent overflows
- Introduced inline wrappers for fseek/ftell/stat with cross-platform handling (Windows, MinGW, Linux) to safely handle large files
- Updated offsets and related variables to larger data types where overflows could occur
- Ensures correct pointer arithmetic and file operations for big data scenarios

+++ LASlib/src/lasreaderbuffered.cpp
@@ -545,7 +545,7 @@ BOOL LASreaderBuffered::copy_point_to_buffer()
     }
     if (buffers != nullptr) 
     {
-      buffers[number_of_buffers] = (U8*)malloc(point.total_point_size * points_per_buffer);
+      buffers[number_of_buffers] = (U8*)malloc((size_t)point.total_point_size * (size_t)points_per_buffer);
       current_buffer = buffers[number_of_buffers];
     }
     number_of_buffers++;
